Travel/travelapp/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.1
Django==4.2.16
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
django-sslserver==0.22
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauth2_provider==0.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
nltk==3.9.1
